test: move them and fix them all
